Moving filter_overlaps-and_sort to utils.py in red_flagger folder.
