+++ data_building/post_annotation_processing.py
@@ -1,10 +1,10 @@
 """A script to run to avoid duplicate words post annotations"""
 import argparse
 import os
-
 import pandas as pd
 
 from red_flagger.obscure_data import obscure, unobscure, ObscuringError
+from red_flagger.utils import filter_overlaps_and_sort
 
 OUT_FILEPATH = os.path.join(os.path.dirname(os.path.dirname(__file__)), "abuse_flagger", "data", "toxic_keywords_b16.txt")
 
@@ -19,7 +19,8 @@ def set_up_parser() -> argparse.ArgumentParser:
         prog='AnnotationPostProcessing',
         description='Processes an annotated wordlist and adds it to the stored global blacklist.'
     )
-    parser.add_argument('filename', help="The path to the annotation file.")
+    #revert before commit
+    parser.add_argument('--filename', help="The path to the annotation file.", default = "./toxic_lexicon_annotated.csv")
     parser.add_argument("--word_col", help="The name of the column containing the toxic words.", default="words_to_check")
     parser.add_argument("--annotation_col", help="The name of the column containing the binary annotations.", default="Abuse Term (Y/N)")
     parser.add_argument("--positive_label", help="The string that represents a positive annotation in the annotation_col.", default="y")
@@ -45,39 +46,6 @@ def validate_data(df: pd.DataFrame, args: argparse.Namespace) -> None:
         raise ArgsError(f"Annotations other than specified positive_label \"{args.positive_label}\" and negative_label \"{args.negative_label}\" found in annotation_col {args.annotation_col}.")
     if os.path.exists(args.out_file):
         input("WARNING!!!!! The out_file exists. Execution of this script will override it.")
-
-
-def filter_overlaps_and_sort(word_list: list[str]) -> list[str]:
-    """Filters the word list to ensure items are unique within the word list.
-
-    Filtering happens in two steps:
-    1) All the single gram items are extracted into a list, iff they are unique to the list. Multi-words are stored.
-    2) The multi-word container is iterated and added to the unique words list iff the multi-words do not contain
-        existing single words.
-
-    This algorithm also pseudo-sorts the list. First, all the unigrams are listed, then the multiword items.
-    """
-    # NOT using a set here, want to be careful about determinism & insertion order.
-    unique_words: list[str] = []
-    multi_words: list[tuple[str, str, ...]] = []
-
-    # gets unigrams 1749 (expected unique) 442 (expected multi)
-    split_word_list = [x.split() for x in word_list]
-    [unique_words.append(x[0]) for x in split_word_list if len(x) == 1 and x[0] not in unique_words]
-    [multi_words.append(x) for x in split_word_list if len(x) > 1 and x not in multi_words]
-
-    # Second loop checks for overlap with longer sequences.
-    for multi_word_tuple in multi_words:
-        # "Are none of the seen words in the current multiword?"
-        if all(
-            [word not in unique_words for word in multi_word_tuple]
-        ):
-            recomposed_multi_word = " ".join(multi_word_tuple)
-            unique_words.append(recomposed_multi_word)
-
-    return unique_words
-
-
 
 if __name__ == "__main__":
     cli_parser = set_up_parser()
